+++ src/components/AboutMe.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { motion } from "framer-motion";
-import { Code, Coffee, Gamepad, Book, Plane, Camera } from "lucide-react";
+import { Code, Coffee, Gamepad, Book, Plane, Camera, Award, Rocket } from "lucide-react";
 
 const AboutMe: React.FC = () => {
   const interests = [
@@ -37,40 +37,29 @@ const AboutMe: React.FC = () => {
   ];
 
   return (
-    <section id="aboutme" className="section bg-gray-800">
-      <div className="container-custom">
-        <motion.div
-          initial={{ opacity: 0, y: 20 }}
-          whileInView={{ opacity: 1, y: 0 }}
-          transition={{ duration: 0.5 }}
-          viewport={{ once: true }}
-          className="mb-12 text-center"
-        >
-          <h2 className="text-3xl font-bold mb-4">More About Me</h2>
-          <div className="w-16 h-1 bg-primary-500 mx-auto"></div>
-          <p className="mt-4 text-gray-300 max-w-2xl mx-auto">
-            Beyond coding, I'm passionate about various interests that shape who I am and influence my creative approach
-            to problem-solving.
+    <section id="about" className="relative py-20 bg-slate-900/50">
+      <div className="max-w-7xl mx-auto px-6">
+        <div className="text-center max-w-3xl mx-auto space-y-6">
+          <h2 className="text-4xl md:text-5xl font-bold">
+            About{" "}
+            <span className="bg-gradient-to-r from-amber-400 to-yellow-500 bg-clip-text text-transparent">Me</span>
+          </h2>
+          <p className="text-lg text-slate-400 leading-relaxed">
+            Frontend developer with experience in creating and optimizing e-commerce applications, specialized in
+            migration strategies, development, and product scalability at leading companies. Passionate about teamwork
+            and continuous learning, I constantly seek new challenges that strengthen my technical skills and contribute
+            to the growth of the projects I participate in.
           </p>
-        </motion.div>
-
-        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
-          {interests.map((interest, index) => (
-            <motion.div
-              key={index}
-              initial={{ opacity: 0, y: 20 }}
-              whileInView={{ opacity: 1, y: 0 }}
-              transition={{ duration: 0.5, delay: index * 0.1 }}
-              viewport={{ once: true }}
-              className="bg-gray-700 rounded-lg p-6 hover:bg-gray-600 transition-colors"
-            >
-              <div className="flex items-center mb-4">
-                <div className="text-primary-400 mr-3">{interest.icon}</div>
-                <h3 className="text-xl font-semibold">{interest.title}</h3>
-              </div>
-              <p className="text-gray-300">{interest.description}</p>
-            </motion.div>
-          ))}
+          <div className="flex flex-wrap justify-center gap-6 pt-4">
+            <div className="flex items-center gap-2 text-slate-300">
+              <Award className="text-amber-400" size={20} />
+              <span>Computer Engineering</span>
+            </div>
+            <div className="flex items-center gap-2 text-slate-300">
+              <Rocket className="text-amber-400" size={20} />
+              <span>4+ Years Experience</span>
+            </div>
+          </div>
         </div>
       </div>
     </section>
