+++ src/components/ContactModal.tsx
@@ -1,15 +1,20 @@
 import React, { useState, useEffect } from "react";
 import { X, Send } from "lucide-react";
 
-function ContactModal({ isOpen, onClose }) {
+interface ContactModalProps {
+  isOpen: boolean;
+  onClose: () => void;
+}
+
+const ContactModal: React.FC<ContactModalProps> = ({ isOpen, onClose }) => {
   const [formData, setFormData] = useState({
     name: "",
     email: "",
     subject: "",
     message: "",
   });
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [submitStatus, setSubmitStatus] = useState(null);
+  const [isSubmitting, setIsSubmitting] = useState<boolean>(false);
+  const [submitStatus, setSubmitStatus] = useState<string | null>(null);
 
   useEffect(() => {
     if (isOpen) {
@@ -22,26 +27,17 @@ function ContactModal({ isOpen, onClose }) {
     };
   }, [isOpen]);
 
-  const handleChange = (e) => {
+  const handleChange = (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
     setFormData({
       ...formData,
       [e.target.name]: e.target.value,
     });
   };
 
-  const handleSubmit = async (e) => {
+  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
     setIsSubmitting(true);
 
-    // TODO: Aquí conectarás tu servicio de email
-    // Ejemplo con fetch a tu API:
-    // const response = await fetch('/api/contact', {
-    //   method: 'POST',
-    //   headers: { 'Content-Type': 'application/json' },
-    //   body: JSON.stringify(formData)
-    // });
-
-    // Por ahora simulamos el envío
     setTimeout(() => {
       console.log("Form data:", formData);
       setSubmitStatus("success");
@@ -60,12 +56,9 @@ function ContactModal({ isOpen, onClose }) {
 
   return (
     <div className="fixed inset-0 z-50 flex items-center justify-center p-4">
-      {/* Backdrop */}
       <div className="absolute inset-0 bg-black/70 backdrop-blur-sm" onClick={onClose}></div>
 
-      {/* Modal */}
       <div className="relative bg-slate-900 border border-slate-700 rounded-2xl w-full max-w-2xl max-h-[90vh] overflow-y-auto shadow-2xl">
-        {/* Header */}
         <div className="sticky top-0 bg-slate-900 border-b border-slate-700 px-6 py-4 flex items-center justify-between">
           <h3 className="text-2xl font-bold">Get in Touch</h3>
           <button onClick={onClose} className="text-slate-400 hover:text-white transition-colors">
@@ -73,7 +66,6 @@ function ContactModal({ isOpen, onClose }) {
           </button>
         </div>
 
-        {/* Form */}
         <form onSubmit={handleSubmit} className="p-6 space-y-6">
           <div className="grid md:grid-cols-2 gap-4">
             <div>
@@ -87,7 +79,7 @@ function ContactModal({ isOpen, onClose }) {
                 value={formData.name}
                 onChange={handleChange}
                 required
-                className="w-full px-4 py-3 bg-slate-800 border border-slate-700 rounded-lg focus:outline-none focus:border-indigo-500 transition-colors text-white"
+                className="w-full px-4 py-3 bg-slate-800 border border-slate-700 rounded-lg focus:outline-none focus:border-amber-500 transition-colors text-white"
                 placeholder="John Doe"
               />
             </div>
@@ -103,7 +95,7 @@ function ContactModal({ isOpen, onClose }) {
                 value={formData.email}
                 onChange={handleChange}
                 required
-                className="w-full px-4 py-3 bg-slate-800 border border-slate-700 rounded-lg focus:outline-none focus:border-indigo-500 transition-colors text-white"
+                className="w-full px-4 py-3 bg-slate-800 border border-slate-700 rounded-lg focus:outline-none focus:border-amber-500 transition-colors text-white"
                 placeholder="john@example.com"
               />
             </div>
@@ -120,7 +112,7 @@ function ContactModal({ isOpen, onClose }) {
               value={formData.subject}
               onChange={handleChange}
               required
-              className="w-full px-4 py-3 bg-slate-800 border border-slate-700 rounded-lg focus:outline-none focus:border-indigo-500 transition-colors text-white"
+              className="w-full px-4 py-3 bg-slate-800 border border-slate-700 rounded-lg focus:outline-none focus:border-amber-500 transition-colors text-white"
               placeholder="Project Opportunity"
             />
           </div>
@@ -135,28 +127,26 @@ function ContactModal({ isOpen, onClose }) {
               value={formData.message}
               onChange={handleChange}
               required
-              rows="6"
-              className="w-full px-4 py-3 bg-slate-800 border border-slate-700 rounded-lg focus:outline-none focus:border-indigo-500 transition-colors text-white resize-none"
+              rows={6}
+              className="w-full px-4 py-3 bg-slate-800 border border-slate-700 rounded-lg focus:outline-none focus:border-amber-500 transition-colors text-white resize-none"
               placeholder="Tell me about your project..."
             ></textarea>
           </div>
 
-          {/* Submit Status */}
           {submitStatus === "success" && (
             <div className="bg-green-500/20 border border-green-500/50 rounded-lg p-4 text-green-400 text-center">
               ✓ Message sent successfully! I'll get back to you soon.
             </div>
           )}
 
-          {/* Submit Button */}
           <button
             type="submit"
             disabled={isSubmitting}
-            className="w-full px-6 py-4 bg-gradient-to-r from-indigo-500 to-purple-600 rounded-lg font-semibold hover:shadow-lg hover:shadow-indigo-500/50 transition-all hover:-translate-y-1 disabled:opacity-50 disabled:cursor-not-allowed disabled:hover:translate-y-0 flex items-center justify-center gap-2"
+            className="w-full px-6 py-4 bg-gradient-to-r from-amber-500 to-yellow-500 text-slate-900 rounded-lg font-semibold hover:shadow-lg hover:shadow-amber-500/50 transition-all hover:-translate-y-1 disabled:opacity-50 disabled:cursor-not-allowed disabled:hover:translate-y-0 flex items-center justify-center gap-2"
           >
             {isSubmitting ? (
               <>
-                <div className="w-5 h-5 border-2 border-white/30 border-t-white rounded-full animate-spin"></div>
+                <div className="w-5 h-5 border-2 border-slate-900/30 border-t-slate-900 rounded-full animate-spin"></div>
                 Sending...
               </>
             ) : (
@@ -170,6 +160,6 @@ function ContactModal({ isOpen, onClose }) {
       </div>
     </div>
   );
-}
+};
 
 export default ContactModal;
