+++ src/components/Navbar.tsx
@@ -1,10 +1,14 @@
 import React, { useState, useEffect } from "react";
-import { Menu } from "lucide-react";
-import { useSidebarStore } from "../store/sidebarStore";
-import logo from "../assets/img/logo.png";
+import { Menu, X } from "lucide-react";
+// import { useSidebarStore } from "../store/sidebarStore";
+import logoImg from "../assets/img/logo.png";
 
 const Navbar: React.FC = () => {
-  const { toggleSidebar } = useSidebarStore();
+  // const { toggleSidebar } = useSidebarStore();
+  const [isMenuOpen, setIsMenuOpen] = useState(false);
+
+  const [activeSection, setActiveSection] = useState("home");
+
   const [scrolled, setScrolled] = useState(false);
 
   useEffect(() => {
@@ -18,40 +22,52 @@ const Navbar: React.FC = () => {
 
   return (
     <header
-      className={`fixed top-0 left-0 right-0 z-50 transition-all duration-300 ${
-        scrolled ? "bg-background-dark/95 backdrop-blur-sm shadow-md py-2" : "bg-transparent py-4"
+      className={`fixed top-0 w-full z-50 transition-all duration-300 ${
+        scrolled ? "bg-slate-900/80 backdrop-blur-lg border-b border-slate-800" : ""
       }`}
     >
-      <nav className="container-custom flex items-center justify-between">
-        <div className="flex items-center">
-          <a href="#home">
-            <img className="w-40" src={logo} />
-          </a>
+      <nav className="max-w-7xl mx-auto px-6 py-4">
+        <div className="flex items-center justify-between">
+          <div className="flex items-center gap-3">
+            <img src={logoImg} alt="Alex Logo" className="h-12" />
+          </div>
+
+          <ul className="hidden md:flex items-center space-x-8">
+            {["home", "about", "experience", "skills", "contact"].map((section) => (
+              <li key={section}>
+                <button
+                  onClick={() => console.log(" section ")}
+                  className={`capitalize hover:text-amber-400 transition-colors ${
+                    activeSection === section ? "text-amber-400" : "text-slate-400"
+                  }`}
+                >
+                  {section}
+                </button>
+              </li>
+            ))}
+          </ul>
+
+          <button onClick={() => setIsMenuOpen(!isMenuOpen)} className="md:hidden text-slate-300 hover:text-white">
+            {isMenuOpen ? <X size={24} /> : <Menu size={24} />}
+          </button>
         </div>
 
-        <div className="hidden md:flex space-x-8">
-          <a href="#home" className="nav-link">
-            Home
-          </a>
-          <a href="#experience" className="nav-link">
-            Experience
-          </a>
-          <a href="#projects" className="nav-link">
-            Projects
-          </a>
-
-          <a href="#aboutme" className="nav-link">
-            About me
-          </a>
-        </div>
-
-        <button
-          onClick={toggleSidebar}
-          className="flex md:hidden text-gray-300 hover:text-white focus:outline-none"
-          aria-label="Toggle menu"
-        >
-          <Menu className="h-6 w-6" />
-        </button>
+        {isMenuOpen && (
+          <div className="md:hidden mt-4 pb-4">
+            <ul className="flex flex-col space-y-4">
+              {["home", "about", "experience", "skills", "contact"].map((section) => (
+                <li key={section}>
+                  <button
+                    onClick={() => console.log(" section ")}
+                    className="capitalize text-slate-300 hover:text-amber-400 transition-colors w-full text-left"
+                  >
+                    {section}
+                  </button>
+                </li>
+              ))}
+            </ul>
+          </div>
+        )}
       </nav>
     </header>
   );
